refa: minor refactor (#7)
* refa(auth): change search services using query params

* refa(auth): solve conflict search user services

* feat(user): add field phone

commit c48c028aaeef3f753698385e36b739f9306c0387
Author: Hafidz Naufal <91860668+hafidznaufl@users.noreply.github.com>
Date:   Tue Nov 14 10:35:28 2023 +0700

    feat(user): add field phone

commit d7b7c5b207a2d5f63550899cdbb53b91ae503f02
Merge: 70bd579 9a3db18
Author: Hafidz Naufal <91860668+hafidznaufl@users.noreply.github.com>
Date:   Mon Nov 13 12:57:26 2023 +0700

    Merge branch 'dev/naufal'  into 'feat/auth'

commit 70bd579d725343c934e1ec1d780c9e70bd1d8bdc
Author: Hafidz Naufal <91860668+hafidznaufl@users.noreply.github.com>
Date:   Sat Nov 11 11:44:46 2023 +0700

    reaf(auth): change using generate id

commit 51cb76edebdcee046b91990235599889b1c2c7e4
Author: Hafidz Naufal <91860668+hafidznaufl@users.noreply.github.com>
Date:   Sat Nov 11 10:39:29 2023 +0700

    refa(admin): change register routing

commit 3a8e69d6be9a9ba494d0d102ce385106895b59c3
Author: Hafidz Naufal <91860668+hafidznaufl@users.noreply.github.com>
Date:   Sat Nov 11 08:19:39 2023 +0700

    refa(user): minor refactor

commit 11cd773e261b3a3df60c919aabb69a987b44bfec
Author: Hafidz Naufal <91860668+hafidznaufl@users.noreply.github.com>
Date:   Thu Nov 9 23:00:28 2023 +0700

    refa(auth): change search services using query params

commit ac581a5bebd6de2ee93746cba92433f005365ee2
Author: Hafidz Naufal <91860668+hafidznaufl@users.noreply.github.com>
Date:   Thu Nov 9 22:40:23 2023 +0700

    refa(auth): change func order by crud

commit 6e5047542ab62d0c8aa821659aea7c383e255840
Author: Hafidz Naufal <91860668+hafidznaufl@users.noreply.github.com>
Date:   Wed Nov 8 21:22:36 2023 +0700

    feat(auth): add auth user & admin services

* refa: minor refactor
