+++ .env.example
@@ -7,3 +7,8 @@ JWT_SECRET=
 JWT_SECRET_ADMIN=
 OPEN_AI_TOKEN=
 FIREBASE_URL=
+SMTP_SERVER=
+SMTP_PORT=
+SMTP_USERNAME=
+SMTP_PASSWORD=
+SENDER_EMAIL=
