+++ service/userService.go
@@ -5,7 +5,7 @@ import (
 	"ecomplaint/model/web"
 	"ecomplaint/repository"
 	"ecomplaint/utils/helper"
-	"ecomplaint/utils/req"
+	req "ecomplaint/utils/request"
 	"fmt"
 
 	"github.com/go-playground/validator"
